Criando hook para o resumo
Fixes #22
